+++ src/Pim/Product/ProductService.php
@@ -7,8 +7,10 @@ declare(strict_types=1);
 
 namespace Wizaplace\SDK\Pim\Product;
 
+use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\RequestOptions;
 use Wizaplace\SDK\AbstractService;
+use Wizaplace\SDK\Exception\NotFound;
 use function theodorejb\polycast\to_int;
 
 final class ProductService extends AbstractService
@@ -16,7 +18,14 @@ final class ProductService extends AbstractService
     public function getProductById(int $productId): Product
     {
         $this->client->mustBeAuthenticated();
-        $data = $this->client->get("products/$productId");
+        try {
+            $data = $this->client->get("products/$productId");
+        } catch (ClientException $e) {
+            if ($e->getCode() === 404) {
+                throw new NotFound("product #${productId} not found", $e);
+            }
+            throw $e;
+        }
 
         return new Product($data);
     }
@@ -50,5 +59,11 @@ final class ProductService extends AbstractService
         ]);
 
         return to_int($data['product_id']);
+    }
+
+    public function deleteProduct(int $productId): void
+    {
+        $this->client->mustBeAuthenticated();
+        $this->client->delete("products/${productId}");
     }
 }